# swagger.yml
swagger: '2.0'
info:
  title: Ethereum Auth0 API
  description: ApT rocks!
  version: "1.0.0"
# the domain of the service
paths:
  /register/user:
    post:
      summary: Registers a new user from the web
      tags: SYNCHRONOUS
      parameters:
        - name: the request
          in: body
          description: registration data
          required: true
          schema:
            $ref: '#/definitions/UserRegistrationRequest'
      responses:
        '201':
          description: Created OK
        '403':
          description: Challenge unsuccesful
          schema:
            $ref: '#/definitions/Error'
  /register/mobile:
    post:
      summary: Registers a new secondary key from the mobile
      tags: ASYNCHRONOUS
      parameters:
        - name: the request
          in: body
          description: registration data
          required: true
          schema:
            $ref: '#/definitions/MobileRegistrationRequest'
      responses:
        '200':
          description: OK
        '403':
          description: Challenge unsuccesful
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  MobileRegistrationRequest:
    type: object
    properties:
      secondaryAddress:
        type: string
        description: Secondary address, format 0x[Aa-fF0-9]{40}
      registrationToken:
        type: string
        description: Data used to identify the device for push notifications.
  UserRegistrationRequest:
    type: object
    properties:
      email:
        type: string
        description: User email.
      primaryAddress:
        type: string
        description: Primary address, format 0x[Aa-fF0-9]{40}
...
  Error:
    type: object
    properties:
      requestId:
        type: string
      message:
        type: string